+++ client/src/main.rs
@@ -1,4 +1,10 @@
+mod game_logic;
+mod interpolator;
+mod render;
+mod ui;
+
 use flatbuffers::FlatBufferBuilder;
+use interpolator::Interpolator;
 use macroquad::math::f32;
 use macroquad::prelude::*;
 use serde::Deserialize;
@@ -7,17 +13,12 @@ use shared::state;
 use shared::state::CommandContent;
 use state::{GameState, PlayerState, PlayerStateCommand};
 use std::collections::{HashMap, VecDeque};
-use std::env::args;
 use std::fs::File;
 use std::net::UdpSocket;
 use std::sync::Arc;
 use std::sync::mpsc::{self, Receiver, Sender};
 use std::time::{Duration, Instant, SystemTime, UNIX_EPOCH};
 use std::{io, thread};
-
-mod game_logic;
-mod render;
-mod ui;
 
 use crate::game_logic::{Screen, UiState};
 use crate::render::render;
@@ -296,64 +297,6 @@ impl Client {
                 }
             }
         }))
-    }
-}
-
-struct Interpolator {
-    old_server_state: GameState,
-    new_server_state: GameState,
-    received_new_state_at: Instant,
-    interpolation_time: Duration,
-    t: f32,
-}
-
-impl Interpolator {
-    fn new(game_state: &GameState) -> Self {
-        Self {
-            old_server_state: game_state.clone(),
-            new_server_state: game_state.clone(),
-            received_new_state_at: Instant::now(),
-            interpolation_time: Duration::ZERO,
-            t: 0.0,
-        }
-    }
-
-    fn set_new_state(&mut self, new_state: GameState) {
-        self.old_server_state = self.new_server_state.clone();
-        self.new_server_state = new_state;
-        self.interpolation_time = Instant::now().duration_since(self.received_new_state_at);
-        self.received_new_state_at = Instant::now();
-    }
-
-    fn interpolate(&mut self, game_state: &mut GameState, client_player_id: u32) {
-        self.update_t();
-        for player in game_state.players.values_mut() {
-            if player.id == client_player_id {
-                continue;
-            }
-
-            let old_position = self
-                .old_server_state
-                .players
-                .get(&player.id)
-                .map_or(state::Vec2::ZERO, |p| p.pos);
-            let new_position = self
-                .new_server_state
-                .players
-                .get(&player.id)
-                .map_or(state::Vec2::ZERO, |p| p.pos);
-
-            player.pos = self.lerp_position(old_position, new_position);
-        }
-    }
-
-    fn update_t(&mut self) {
-        self.t = self.received_new_state_at.elapsed().as_millis() as f32
-            / self.interpolation_time.as_millis() as f32;
-    }
-
-    fn lerp_position(&self, start: state::Vec2, end: state::Vec2) -> state::Vec2 {
-        start + (end - start) * self.t
     }
 }
